Finished Determine Accuracy Parameters and saw no impact
It does not look like considering the most frequent 1000 to 5000 words affects the predictive value at all. Also it does not seem like stop words affect things. The take away is that you should strip stopwords( since this will make it faster) and just pick arbitrarily 4000 is a nice round number the number of features to consider.

The next relevant thing to test is different parts of speech to see if that affects things

Look at AlgoEvaluationResutls.txt for the different combinations.